Switched for a completely cssharp solution (no signatures required)
